Saving Works w/ Character | Read the des.
Loading works, but without the character. The problem is injecting the
character into the class while updating the grid to match the class.
Ahmed has created it so it only checks once setting up the grid. It's
going to require an if statement, will check with you guys tomorrow.
Back-end works, its just a matter of adding it to the GUI.